Update 90 percent_CNN GAN1.py
This code is a python script that uses various libraries like Numpy, Tensorflow, OpenCV, Keras etc. to create, train and test a Convolutional Neural Network (CNN) for image classification. The code reads and loads the input spectrum images from a specified folder path and pre-processes the images by resizing and normalizing. It then uses the Keras library to define, compile and train the CNN model. The code also implements early stopping, checkpointing and 10-fold cross-validation to prevent overfitting and improve the accuracy of the model